bank tags: fix npe on shutdown (#17759)

+++ runelite-client/src/main/java/net/runelite/client/plugins/banktags/tabs/TabInterface.java
@@ -394,8 +394,12 @@ public class TabInterface
 		plugin.open(null);
 
 		upButton = downButton = newTab = scrollComponent = null;
-		parent.deleteAllChildren();
-		parent = null;
+
+		if (parent != null)
+		{
+			parent.deleteAllChildren();
+			parent = null;
+		}
 
 		tabManager.clear();
 	}
